analytics/requirements.txt
pandas==2.0.0
matplotlib==3.7.1
requests==2.28.2
schedule==1.1.0
python-dotenv==1.0.0
google-api-python-client==2.86.0
oauth2client==4.1.3 